Change reverse_tolerance type to uint32_t

+++ src/conf/datatypes.h
@@ -229,7 +229,7 @@ typedef struct {
     bool enable_quickstop;
     bool fault_darkride_enabled;
     bool fault_reversestop_enabled;
-    float reverse_tolerance;
+    uint32_t reverse_tolerance;
     float tiltback_duty_angle;
     float tiltback_duty_speed;
     float tiltback_duty;
